refactor navigate function in its own hook
