+++ src/util.ts
@@ -1,5 +1,9 @@
-import {useState} from 'react';
+import {useState, useTransition} from 'react';
 import {ILocation} from './types';
+// @ts-ignore
+import {createFromReadableStream} from 'react-server-dom-webpack';
+import {useRefresh} from './Cache.client';
+import {useLocation} from './LocationContext.client';
 
 export function useMutation({
   endpoint,
@@ -47,4 +51,27 @@ export function useMutation({
   }
 
   return {isSaving, performMutation};
+}
+
+export function useNavigation() {
+  const refresh = useRefresh();
+  const [isNavigating, startNavigating] = useTransition();
+  const {location, setLocation} = useLocation();
+
+  function navigate(response: Response) {
+    const cacheKey = response.headers.get('X-Location');
+
+    if (!cacheKey) {
+      throw new Error('X-Location header is not set');
+    }
+
+    const nextLocation = JSON.parse(cacheKey);
+    const seededResponse = createFromReadableStream(response.body);
+    startNavigating(() => {
+      refresh(cacheKey, seededResponse);
+      setLocation && setLocation(nextLocation);
+    });
+  }
+
+  return {isNavigating, location, navigate};
 }
